♻️ refactor(repository): move AdviceCompletionDto to dto package
Relocate AdviceCompletionDto from entities to a dedicated dto package
to improve project structure and separation of concerns.
