+++ app/src/main/java/com/akilimo/mobile/dto/AdviceCompletionDto.kt
@@ -1,0 +1,12 @@
+package com.akilimo.mobile.dto
+
+import android.os.Parcelable
+import com.akilimo.mobile.enums.EnumAdviceTask
+import com.akilimo.mobile.enums.EnumStepStatus
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class AdviceCompletionDto(
+    val taskName: EnumAdviceTask,
+    val stepStatus: EnumStepStatus
+) : Parcelable
