+++ app/src/main/java/com/akilimo/mobile/repos/AdviceCompletionRepo.kt
@@ -2,7 +2,7 @@ package com.akilimo.mobile.repos
 
 import com.akilimo.mobile.dao.AdviceCompletionDao
 import com.akilimo.mobile.entities.AdviceCompletion
-import com.akilimo.mobile.entities.AdviceCompletionDto
+import com.akilimo.mobile.dto.AdviceCompletionDto
 import com.akilimo.mobile.enums.EnumAdviceTask
 import com.akilimo.mobile.enums.EnumStepStatus
 import kotlinx.coroutines.flow.Flow
